Abstraindo componentes.
