+++ web/src/App.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 
 import './styles/global.css';
+import 'leaflet/dist/leaflet.css';
 
 import Routes from './routes';
