+++ web/src/pages/CreateOrphanage.tsx
@@ -1,37 +1,16 @@
 import React from "react";
 import { Map, Marker, TileLayer } from 'react-leaflet';
-import L from 'leaflet';
-import { useHistory } from "react-router-dom";
+import { FiPlus } from "react-icons/fi";
 
-import { FiArrowLeft, FiPlus } from "react-icons/fi";
-
-import mapMarkerImg from '../images/map-marker.svg';
+import mapIcon from "../utils/mapIcon";
+import Sidebar from "../componentes/sidebar";
 
 import '../styles/pages/create-orphanage.css';
 
-const happyMapIcon = L.icon({
-  iconUrl: mapMarkerImg,
-
-  iconSize: [58, 68],
-  iconAnchor: [29, 68],
-  popupAnchor: [0, -60]
-})
-
 export default function CreateOrphanage() {
-  const { goBack } = useHistory();
-
   return (
     <div id="page-create-orphanage">
-      <aside>
-        <img src={mapMarkerImg} alt="Happy" />
-
-        <footer>
-          <button type="button" onClick={goBack}>
-            <FiArrowLeft size={24} color="#FFF" />
-          </button>
-        </footer>
-      </aside>
-
+      <Sidebar />
       <main>
         <form className="create-orphanage-form">
           <fieldset>
@@ -42,11 +21,9 @@ export default function CreateOrphanage() {
               style={{ width: '100%', height: 280 }}
               zoom={15}
             >
-              <TileLayer 
-                url={`https://api.mapbox.com/styles/v1/mapbox/light-v10/tiles/256/{z}/{x}/{y}@2x?access_token=${process.env.REACT_APP_MAPBOX_TOKEN}`}
-              />
+              <TileLayer url='https://a.tile.openstreetmap.org/{z}/{x}/{y}.png' />
 
-              <Marker interactive={false} icon={happyMapIcon} position={[-27.2092052,-49.6401092]} />
+              <Marker interactive={false} icon={mapIcon} position={[-27.2092052,-49.6401092]} />
             </Map>
 
             <div className="input-block">
